legacy_django/requirements.txt
Django>=4.2
psycopg2-binary
Pillow
django-crispy-forms
crispy-bootstrap5
dj-database-url
whitenoise
gunicorn
djangorestframework
django-cors-headers
